requirements.txt
PyJWT>=2.9.0,<3.0.0
